+++ src/app/modules/blog/post-card/post-card.component.html
